+++ tests.sh
@@ -78,7 +78,7 @@ done
 echo '-> 2 Test plugins installation with tar.bz2'
 image_name=glpi_2
 docker run $docker_run_options --name "${image_name}" --env='GLPI_INSTALL_PLUGINS=fusioninventory|https://github.com/fusioninventory/fusioninventory-for-glpi/releases/download/glpi9.3%2B1.1/fusioninventory-9.3.1.1.tar.bz2' "${image_building_name}"
-wait_for_string_in_container_logs "${image_name}" 'Starting up...'
+wait_for_string_in_container_logs "${image_name}" 'GET /fpm-ping'
 # test
 if ! docker exec "${image_name}" test -d plugins/fusioninventory; then
   docker logs "${image_name}"
@@ -91,7 +91,7 @@ stop_and_remove_container "${image_name}"
 echo '-> 3 Test plugins installation with tar.gz'
 image_name=glpi_3
 docker run $docker_run_options --name "${image_name}" --env='GLPI_INSTALL_PLUGINS=fusioninventory|https://github.com/fusioninventory/fusioninventory-for-glpi/releases/download/glpi9.3%2B1.2/fusioninventory-9.3+1.2.tar.gz' "${image_building_name}"
-wait_for_string_in_container_logs "${image_name}" 'Starting up...'
+wait_for_string_in_container_logs "${image_name}" 'GET /fpm-ping'
 # test
 if ! docker exec "${image_name}" test -d plugins/fusioninventory; then
   docker logs "${image_name}"
@@ -118,7 +118,7 @@ stop_and_remove_container "${image_name}"
 echo '-> 6 Test plugins installation with zip'
 image_name=glpi_6
 docker run $docker_run_options --name "${image_name}" --env='GLPI_INSTALL_PLUGINS=timezones|https://github.com/tomolimo/timezones/releases/download/2.4.1/timezones-2.4.1.zip' "${image_building_name}"
-wait_for_string_in_container_logs "${image_name}" 'Starting up...'
+wait_for_string_in_container_logs "${image_name}" 'GET /fpm-ping'
 # test
 if ! docker exec "${image_name}" test -d plugins/timezones; then
   docker logs "${image_name}"